Rolling Code Trigger Implemented
Added additional colors for the Display.  When the system
is activated, the display goes from blue to Violet.

Timing tweaked for my setup, to get smooth transitions from 2->3 and
then back to 2.

Note that the speed of the center leg drop is reduced to half speed.
At full speed, it misses the limit switch and blows right through.