Moving the 3->2 transition to button D.
In the current revision of the PCB, Button C is always reading HIGH,
meaning the code thinks that the button is always pressed.  This
results in the 2->3 transition failing immediately, as the code
thinks that the 3->2 button is being pressed.

As soon as the killswitch (button A) is pressed, this insta triggers
the 3->2 transition .... THIS IS BAD!

The workaround for now, is to change the 3->2 button from button C
to button D, and ignore button D.  Not a big issue as the D button was
not used.

I'll investigate the board, and get thigns fixed.  Fot now though the
sketch will work with version 1.0 of the PCB.